+++ examples/test/defaults/test_defaults_bad_group.pas
@@ -1,0 +1,7 @@
+program TestDefaultsBadGroup;
+procedure P(x, y: integer = 5);
+begin
+end;
+
+begin
+end.