+++ examples/test/defaults/test_defaults_bad_nonconst.pas
@@ -1,0 +1,8 @@
+program TestDefaultsBadNonconst;
+var g: integer;
+procedure P(x: integer = g);
+begin
+end;
+
+begin
+end.