+++ examples/test/defaults/test_defaults_bad_nontrailing.pas
@@ -1,0 +1,7 @@
+program TestDefaultsBadNontrailing;
+procedure P(x: integer = 1; y: integer);
+begin
+end;
+
+begin
+end.