+++ examples/test/defaults/test_defaults_bad_out.pas
@@ -1,0 +1,7 @@
+program TestDefaultsBadOut;
+procedure P(out x: integer = 5);
+begin
+end;
+
+begin
+end.