+++ examples/test/defaults/test_defaults_bad_proctype.pas
@@ -1,0 +1,13 @@
+program TestDefaultsBadProctype;
+function Square(n: integer): integer;
+begin
+    Square := n * n;
+end;
+
+function Apply(function f(n: integer): integer = Square; v: integer): integer;
+begin
+    Apply := f(v);
+end;
+
+begin
+end.