+++ examples/test/defaults/test_defaults_bad_record.pas
@@ -1,0 +1,8 @@
+program TestDefaultsBadRecord;
+type TPoint = record x, y: integer; end;
+procedure P(p: TPoint = 5);
+begin
+end;
+
+begin
+end.