+++ examples/test/defaults/test_defaults_bad_subrange.pas
@@ -1,0 +1,8 @@
+program TestDefaultsBadSubrange;
+type TRange = 1..10;
+procedure P(x: TRange = 5);
+begin
+end;
+
+begin
+end.