+++ examples/test/defaults/test_defaults_bad_toofew.pas
@@ -1,0 +1,8 @@
+program TestDefaultsBadToofew;
+procedure P(a: integer; b: integer; c: integer = 1);
+begin
+end;
+
+begin
+    P(1);
+end.