+++ examples/test/defaults/test_defaults_bad_toomany.pas
@@ -1,0 +1,8 @@
+program TestDefaultsBadToomany;
+procedure P(a: integer; b: integer = 2);
+begin
+end;
+
+begin
+    P(1, 2, 3);
+end.