+++ examples/test/defaults/test_defaults_bad_typemismatch.pas
@@ -1,0 +1,7 @@
+program TestDefaultsBadTypemismatch;
+procedure P(x: boolean = 5);
+begin
+end;
+
+begin
+end.