+++ examples/test/defaults/test_defaults_bad_var.pas
@@ -1,0 +1,7 @@
+program TestDefaultsBadVar;
+procedure P(var x: integer = 5);
+begin
+end;
+
+begin
+end.